chapter_002/src/test/java/ru/job4j/comparator/StringComparatorTest.java
package ru.job4j.comparator;
import org.junit.Test;

import static org.hamcrest.core.Is.is;
import static org.hamcrest.Matchers.greaterThan;
import static org.hamcrest.Matchers.lessThan;
import static org.hamcrest.MatcherAssert.assertThat;

/**
 * Класс для тестирования StringComparator
 */
public class StringComparatorTest {
    @Test
    public void whenStringsAreEqualThenZero() {
        StringComparator compare = new StringComparator();
        int rst = compare.compare("Ivanov",
                "Ivanov");
        assertThat(rst, is(0));
    }

    @Test
    public void whenLeftLessThanRightResultShouldBeNegative() {
        StringComparator compare = new StringComparator();
        int rst = compare.compare("Ivanov",
                "Ivanova");
        assertThat(rst, lessThan(0));
    }

    @Test
    public void whenLeftGreaterThanRightResultShouldBePositive() {
        StringComparator compare = new StringComparator();
        int rst = compare.compare("Petrov",
                "Ivanova");
        assertThat(rst, greaterThan(0));
    }

    @Test
    public void secondCharOfLeftGreaterThanRightShouldBePositive() {
        StringComparator compare = new StringComparator();
        int rst = compare.compare("Petrov",
                "Patrov");
        assertThat(rst, greaterThan(0));
    }

    @Test
    public void secondCharOfLeftLessThanRightShouldBeNegative() {
        StringComparator compare = new StringComparator();
        int rst = compare.compare("Patrova",
                "Petrov");
        assertThat(rst, lessThan(0));
    }
}